few changes on collisions

+++ src/Domain/Settings.java
@@ -35,7 +35,7 @@ public class Settings {
 	private int zetaNumber;
 	
 	private Dimension screenSize;
-	public double timeRemaining;
+	public int timeRemaining;
 	
 	public static int timeMult = 20;
